feat(storybook): Add light/dark theme support
Storybook nutzt jetzt @storybook/addon-themes um zwischen dem light und dark Theme zu wechseln. Das Story Element in Storybook wird vom ThemeProvider aus @northware/ui umschlossen. So wird das dark Theme laut der global.css angewendet. Aktuell lässt sich noch nicht stuern, welchen Wert die Primary-Klassen haben (@northware/ui theme-* Klassen)
