+++ apps/storybook/.storybook/main.ts
@@ -1,14 +1,14 @@
 import type { StorybookConfig } from "@storybook/nextjs";
 
-import { dirname, join } from "path";
+import { dirname, join } from "node:path";
 
 /**
  * This function is used to resolve the absolute path of a package.
  * It is needed in projects that use Yarn PnP or are set up within a monorepo.
  */
-function getAbsolutePath(value: string): any {
-  return dirname(require.resolve(join(value, "package.json")));
-}
+const getAbsolutePath = (value: string) =>
+  dirname(require.resolve(join(value, "package.json")));
+
 const config: StorybookConfig = {
   stories: [
     "../stories/**/*.mdx",
@@ -17,6 +17,7 @@ const config: StorybookConfig = {
   addons: [
     getAbsolutePath("@storybook/addon-essentials"),
     getAbsolutePath("@chromatic-com/storybook"),
+    getAbsolutePath("@storybook/addon-themes"),
   ],
   framework: {
     name: getAbsolutePath("@storybook/nextjs"),
