+++ apps/storybook/.storybook/preview.tsx
@@ -1,0 +1,36 @@
+import { ThemeProvider } from "@northware/ui/components/providers/theme-provider";
+import { withThemeByClassName } from "@storybook/addon-themes";
+import type { Preview } from "@storybook/react";
+// biome-ignore lint/correctness/noUnusedImports: React has to be imported to use jsx in this file.
+import React from "react";
+
+import "@northware/ui/css";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    withThemeByClassName({
+      themes: {
+        light: "light",
+        dark: "dark",
+      },
+      defaultTheme: "light",
+    }),
+    (Story) => (
+      <div className="bg-background">
+        <ThemeProvider>
+          <Story />
+        </ThemeProvider>
+      </div>
+    ),
+  ],
+};
+
+export default preview;
