+++ apps/storybook/package.json
@@ -19,6 +19,7 @@
   "devDependencies": {
     "@chromatic-com/storybook": "3",
     "@storybook/addon-essentials": "8.6.10",
+    "@storybook/addon-themes": "^8.6.10",
     "@storybook/blocks": "8.6.10",
     "@storybook/nextjs": "8.6.10",
     "@storybook/react": "8.6.10",
@@ -30,10 +31,5 @@
     "storybook": "8.6.10",
     "tailwindcss": "^4",
     "typescript": "^5"
-  },
-  "eslintConfig": {
-    "extends": [
-      "plugin:storybook/recommended"
-    ]
   }
 }
